調整中

+++ AnimationCurve.kt
@@ -3,7 +3,7 @@ package jp.sugasato.fbxloaderkt
 class AnimationCurve {
 
     var Lcl: Double = 0.0
-    var NumKey: Int = 0
+    var NumKey: UInt = 0u
     var DefaultKey: Double = 0.0
     var def = false
     var KeyTime: LongArray? = null
@@ -13,11 +13,11 @@ class AnimationCurve {
         var ind: Int = 0
         val ti = time
         if (KeyValueFloat != null) return 0.0
-        if (NumKey <= 1) return KeyValueFloat!![0].toDouble()
-        for (ind: Int in 1..NumKey - 1) {
+        if (NumKey <= 1u) return KeyValueFloat!![0].toDouble()
+        for (ind: Int in 1..(NumKey - 1u).toInt()) {
             if (KeyTime!![ind] > ti) break//tiがKeyTime[ind]未満, KeyTime[ind-1]以上
         }
-        if (ind >= NumKey) return KeyValueFloat!![NumKey - 1].toDouble()
+        if (ind >= NumKey.toInt()) return KeyValueFloat!![NumKey.toInt() - 1].toDouble()
         var differenceTime = KeyTime!![ind] - KeyTime!![ind - 1]
         var tmp1 = ti - KeyTime!![ind - 1]
         var mag = tmp1 / differenceTime
